# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XSecurity-Vault",
    "description": "Secure digital certificate management orchestrator",
    "version": "2.0.0"
  },
  "paths": {
...
    "/grupos/empresa/{empresa_id}": {
      "get": {
        "tags": [
          "Grupos"
        ],
        "summary": "Listar Grupos Empresa",
        "operationId": "listar_grupos_empresa_grupos_empresa__empresa_id__get",
        "parameters": [
          {
            "name": "empresa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empresa Id"
            }
          },
          {
            "name": "plano_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plano Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}": {
      "get": {
        "tags": [
          "Grupos"
        ],
        "summary": "Obter",
        "operationId": "obter_grupos__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Grupos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_grupos__grupo_id__put",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Grupos"
        ],
        "summary": "Remover",
        "operationId": "remover_grupos__grupo_id__delete",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/certificados": {
      "get": {
        "tags": [
          "Grupos"
        ],
        "summary": "Listar Certificados Grupo",
        "operationId": "listar_certificados_grupo_grupos__grupo_id__certificados_get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Adicionar Certificado",
        "operationId": "adicionar_certificado_grupos__grupo_id__certificados_post",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/": {
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Criar",
        "operationId": "criar_grupos__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/usuarios": {
      "get": {
        "tags": [
          "Grupos"
        ],
        "summary": "Listar Usuarios Do Grupo",
        "operationId": "listar_usuarios_do_grupo_grupos__grupo_id__usuarios_get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Adicionar Usuario Ao Grupo",
        "operationId": "adicionar_usuario_ao_grupo_grupos__grupo_id__usuarios_post",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/usuarios/{usuario_id}": {
      "delete": {
        "tags": [
          "Grupos"
        ],
        "summary": "Remover Usuario Do Grupo",
        "operationId": "remover_usuario_do_grupo_grupos__grupo_id__usuarios__usuario_id__delete",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/usuarios/bulk": {
      "post": {
        "tags": [
          "Grupos"
        ],
        "summary": "Adicionar Usuarios Em Lote",
        "operationId": "adicionar_usuarios_em_lote_grupos__grupo_id__usuarios_bulk_post",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrupoUsuarioBulkCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/grupos/{grupo_id}/remover/certificado": {
      "delete": {
        "tags": [
          "Grupos"
        ],
        "summary": "Remover Certificado",
        "operationId": "remover_certificado_grupos__grupo_id__remover_certificado_delete",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/grupos-certificados/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Grupos Certificados"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_grupos_certificados_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrupoCertOut"
                  },
                  "title": "Response Listar Grupo Grupos Certificados Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/grupos-usuarios/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Grupos Usu\u00e1rios"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_grupos_usuarios_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrupoUsuarioOut"
                  },
                  "title": "Response Listar Grupo Grupos Usuarios Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/regras-acesso/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_regras_acesso_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoOut"
                  },
                  "title": "Response Listar Grupo Regras Acesso Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/regras-acesso-hosts/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_regras_acesso_hosts_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoHostOut"
                  },
                  "title": "Response Listar Grupo Regras Acesso Hosts Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GrupoCertOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "grupo_cert_id": {
            "type": "integer",
            "title": "Grupo Cert Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "grupo_cert_id"
        ],
        "title": "GrupoCertOut"
      },
...
      "GrupoUsuarioBulkCreate": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "usuario_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Usuario Ids"
          },
          "empresa_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "usuario_ids"
        ],
        "title": "GrupoUsuarioBulkCreate"
      },
...
      "GrupoUsuarioOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "grupo_usuario_id": {
            "type": "integer",
            "title": "Grupo Usuario Id"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "grupo_usuario_id"
        ],
        "title": "GrupoUsuarioOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegraAcessoHostOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          },
          "urls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urls"
          },
          "bloquear_em_feriado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloquear Em Feriado",
            "default": false
          },
          "regra_id": {
            "type": "integer",
            "title": "Regra Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios",
          "regra_id",
          "criado_em"
        ],
        "title": "RegraAcessoHostOut"
      },
...
      "RegraAcessoOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          },
          "regra_id": {
            "type": "integer",
            "title": "Regra Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios",
          "regra_id",
          "criado_em"
        ],
        "title": "RegraAcessoOut"
      },
...
      "TipoDiaEnum": {
        "type": "string",
        "enum": [
          "corridos",
          "uteis",
          "especificos"
        ],
        "title": "TipoDiaEnum"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Session token for authentication. Use the session_token from login response.",
        "scheme": "bearer"
      }
    }
  }
}